changes values for system to test headlights
